+++ test/Fixture/Expressions/NoAssignByReferenceRule/script.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Ergebnis\PHPStan\Rules\Test\Fixture\Expressions\NoAssignByReferenceRule;
+
+$foo = 'bar';
+
+$bar = $foo;
+
+$baz = &$foo;