+++ CHANGELOG.md
@@ -11,6 +11,7 @@ For a full diff see [`2.7.0...main`][2.7.0...main].
 ### Added
 
 - Added `allRules` parameter to allow disabling and enabling all rules  ([#913]), by [@localheinz]
+- Added `Expressions\NoAssignByReferenceRule`, which reports an error when a variable is assigned by reference ([#914]), by [@localheinz]
 
 ## [`2.7.0`][2.7.0]
 
@@ -617,6 +618,7 @@ For a full diff see [`362c7ea...0.1.0`][362c7ea...0.1.0].
 [#911]: https://github.com/ergebnis/phpstan-rules/pull/911
 [#912]: https://github.com/ergebnis/phpstan-rules/pull/912
 [#913]: https://github.com/ergebnis/phpstan-rules/pull/913
+[#914]: https://github.com/ergebnis/phpstan-rules/pull/914
 
 [@cosmastech]: https://github.com/cosmastech
 [@enumag]: https://github.com/enumag